Initial attempt at 301 redirecting.
It would work and probably does for some posts, except for a bug
in the date parsing that I had at first. Also, there are problems
with names like "PowerShell" getting turned into "power_shell" -
I'm not sure of the best way to handle this yet, short of more
metadata during the conversion process.