Add client-side support for lang identifier.
The general idea is the same as what is found in GitHub Flavored
Markdown, except that this code uses highlight.js and a small amount of
javascript.

Above the code block to be highlighted, just add a ```lang identifier,
like ```cs for it to be marked as C#.

The implementation ultimately calls hljs.highlight(lang, code_block).